% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{anchor_features}
\alias{anchor_features}
\title{Find integration anchors and set variable features}
\usage{
anchor_features(object, split.by = "Patient", nfeatures = 2000)
}
\arguments{
\item{object}{A Seurat object.}

\item{split.by}{Character string specifying the metadata column to split by.
Default is "Patient"}

\item{nfeatures}{Number of features to use as integration anchors. Default
uses the function parameter nfeatures}
}
\value{
A Seurat object with anchor features set as variable features
}
\description{
This function aims to use anchor features as variable features for the object.
}
\examples{
\dontrun{
seurat_obj <- anchor_features(seurat_obj, split.by = "Patient", nfeatures = 2000)
}

}
