% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioYourOwnBowl-package.R
\docType{package}
\name{BioYourOwnBowl-package}
\alias{BioYourOwnBowl}
\alias{BioYourOwnBowl-package}
\title{BioYourOwnBowl: Custom Seurat Utilities for Single-Cell Analysis}
\description{
Provides custom utility functions for Seurat single-cell RNA sequencing data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/june-zhang-bioinfo/BioYourOwnBowl}
  \item Report bugs at \url{https://github.com/june-zhang-bioinfo/BioYourOwnBowl/issues}
}

}
\author{
\strong{Maintainer}: June Zhang \email{zhang24925@gmail.com}

}
\keyword{internal}
