% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_vis.R
\name{cnmf_bar_plots}
\alias{cnmf_bar_plots}
\title{Plot Top N Usage Columns From cNMF Gene Spectra}
\usage{
cnmf_bar_plots(
  dir_path,
  object,
  output_dir,
  top_n = 30,
  show_metrics = FALSE,
  highlight_genes = NULL
)
}
\arguments{
\item{dir_path}{Character. Directory containing cNMF output files.}

\item{object}{Seurat or SingleCellExperiment object containing normalized data.}

\item{output_dir}{Directory where PNG plots should be saved.}

\item{top_n}{Number of top genes to extract and plot (default = 30).}

\item{show_metrics}{Logical. If TRUE, add min/median/max labels to bars.}

\item{highlight_genes}{A gene list to be bold and red on barplots. If NULL, output without hightlight.}
}
\value{
A list containing:
\item{top_df}{Data frame of top-N genes per program}
\item{file_used}{Path to the CNMF spectrum file used}
}
\description{
This function reads cNMF gene spectra, normalizes them, extracts gene-level
statistics (min/median/max) from a Seurat object, and produces barplots for
the top-N contributing genes per program. It also returns a top-gene table.
}
