% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_vis.R
\name{cnmf_gene_ranks}
\alias{cnmf_gene_ranks}
\title{Search Gene Ranks Across cNMF Programs}
\usage{
cnmf_gene_ranks(dir_path, features)
}
\arguments{
\item{dir_path}{Path to the directory containing cNMF output files}

\item{features}{Character vector of gene names to search for}
}
\value{
A data frame with genes as rows and programs as columns, showing ranks
}
\description{
Search Gene Ranks Across cNMF Programs
}
