% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_vis.R
\name{dimplots_split}
\alias{dimplots_split}
\title{Save Split DimPlots from a Seurat Object}
\usage{
dimplots_split(object, split_var, group_var, colors, file_name = NULL)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{split_var}{A character string specifying the metadata column to split plots by.}

\item{group_var}{A character string specifying the metadata column to color (group) cells by.}

\item{colors}{A character vector of colors to use for the groups.}

\item{file_name}{Character string for output png name. If NULL, plot will be displayed but not saved.
Default is NULL.}
}
\value{
Invisibly returns the file path of the saved plot.
}
\description{
This function generates and saves UMAP/tSNE DimPlots from a Seurat object,
split by a specified metadata variable. The plots are arranged automatically
based on the number of levels in the splitting variable.
}
\examples{
\dontrun{
batch_vars <- "Cohort"
my_colors <- c("red", "blue", "green")
dimplots_split(
  object = object,
  split_var = batch_vars,
  group_var = "seurat_clusters",
  colors = my_colors,
  file_name = NULL
)
}

}
