% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{dotplots_pdf}
\alias{dotplots_pdf}
\title{Generate DotPlot PDFs for a Seurat Object}
\usage{
dotplots_pdf(object, features, file_name = "dotplot.pdf")
}
\arguments{
\item{object}{A Seurat object.}

\item{file_name}{A string giving the output pdf filename (\code{dotplot.pdf}).}

\item{features.1}{A character vector of feature names to include in the first DotPlot.}

\item{features.2}{A character vector of feature names to include in the second DotPlot.}
}
\value{
A list containing:
\describe{
\item{markers}{The full marker table from FindAllMarkers()}
\item{genes}{The selected marker genes returned by select_marker_genes_score()}
}
}
\description{
This function creates a PDF containing two DotPlots for a given Seurat object:
one using a set of selected features, and another using top marker genes identified
by \code{FindAllMarkers()} and selected with \code{select_marker_genes_score()}.
}
\examples{
\dontrun{
dotplots_pdf(object = seurat_obj,
                     features.1 = c("GeneA", "GeneB"),
                     features.2 = c("GeneC", "GeneD"),
                     file_name = "dotplot.pdf")
}
}
