% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{dotplots_png}
\alias{dotplots_png}
\title{Generate and save a Seurat DotPlot dynamically}
\usage{
dotplots_png(
  object,
  features,
  idents,
  file_name,
  width_factor = 25,
  height_factor = 42,
  res = 100,
  low_color = "#0024d6",
  mid_color = "#b4b6bf",
  high_color = "#d91111",
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{features}{A gene list.}

\item{idents}{The identity column to group cells. E.g. seurat_clusters.}

\item{file_name}{Output file name (e.g., "dotplot.png").}

\item{width_factor}{Width multiplier per gene (default: 25).}

\item{height_factor}{Height multiplier per cluster (default: 40).}

\item{res}{Plot resolution in dpi (default: 100).}

\item{low_color, mid_color, high_color}{Colors for the expression gradient.}
}
\value{
Saves a PNG file of the dot plot.
}
\description{
Generate and save a Seurat DotPlot dynamically
}
