% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{filter_genes_by_expression}
\alias{filter_genes_by_expression}
\title{Filter Genes by Minimum Cell Expression}
\usage{
filter_genes_by_expression(object, gene_filter = 50, assay = "RNA")
}
\arguments{
\item{object}{A Seurat object.}

\item{gene_filter}{Integer. Minimum number of cells a gene must be expressed in to be kept. Default is 50.}

\item{assay}{Character. Name of the assay to use for filtering. Default is "RNA".}
}
\value{
A Seurat object containing only the genes that meet the expression threshold.
}
\description{
This function filters genes in a Seurat object based on the number of cells
in which they are expressed. Only genes expressed in at least \code{gene_filter}
cells are retained.
}
