% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{remove_low_quality_clusters}
\alias{remove_low_quality_clusters}
\title{Remove low-quality clusters based on feature count thresholds}
\usage{
remove_low_quality_clusters(
  object,
  lower_bound_param = 1,
  higher_bound_param = 2
)
}
\arguments{
\item{object}{A Seurat object.}

\item{lower_bound_param}{Numeric; the number of standard deviations below the mean
nFeature_RNA that a cluster must fall to be considered low-quality. To be more stringent, decrease this parameter.
Default is 1.}

\item{higher_bound_param}{Numeric; the number of standard deviations above the mean
nFeature_RNA that a cluster must exceed to be considered low-quality. To be more stringent, decrease this parameter.
Default is 2.}
}
\value{
A Seurat object with low-quality clusters removed.
}
\description{
Remove low-quality clusters based on feature count thresholds
}
