% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{select_marker_genes_rank}
\alias{select_marker_genes_rank}
\title{Select marker genes based on p-value or log2FC ranking}
\usage{
select_marker_genes_rank(
  markers,
  top_n = 25,
  direction = "up",
  rank_by = "p_val_adj",
  adj_p_cutoff = 0.05,
  log2fc_cutoff = 0
)
}
\arguments{
\item{markers}{A data frame or tibble containing marker gene statistics.
Must include the following columns:
\describe{
\item{gene}{Gene name or identifier.}
\item{cluster}{Cluster or group label.}
\item{p_val_adj}{Adjusted p-value (typically from a Seurat differential expression test).}
\item{avg_log2FC}{Average log2 fold change between the cluster and others.}
}}

\item{top_n}{Integer specifying the number of top marker genes to select per cluster,
ranked by score. Default is 25.}

\item{direction}{Character string specifying whether to select upregulated
(\code{"up"}) or downregulated (\code{"down"}) genes. Default is \code{"up"}.}

\item{rank_by}{Character string specifying ranking metric:
"p_val_adj" or "avg_log2FC". Default is p_val_adj.}

\item{adj_p_cutoff}{Numeric value specifying the statistical significance cutoff.}

\item{log2fc_cutoff}{Numeric value specifying the minimum absolute log2 fold change
threshold. Default is 0.}
}
\value{
A data frame of selected marker genes.
}
\description{
This function selects marker genes from a differential expression results table
by ranking either statistical significance (\code{p_val_adj}) or
effect size (\code{avg_log2FC}). It can take results from both \code{FindAllMarkers} (> 2 groups) and \code{FindMarkers} (2 groups).
}
\details{
If using results from \code{FindMarkers}, \code{direction} doesn't take effect. Please make sure you add \code{cluster} column indicating the group name in the input.
}
