% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_vis.R
\name{density_plot}
\alias{density_plot}
\title{Generate Python Code for Scanpy UMAP Density Plots}
\usage{
density_plot(file_name, density_specs, base_figsize = c(4, 4), max_per_row = 4)
}
\arguments{
\item{file_name}{Character string. Path to the \code{.h5ad} file that contains
the AnnData object to be plotted.}

\item{density_specs}{A named list specifying plotting order for each metadata
variable. Each element should itself be a list containing an \code{order}
vector (e.g., \code{list(order = c("group1", "group2", ...))}).}

\item{base_figsize}{Numeric vector of length 2 giving the base figure size
in inches for each density plot panel. Default is \code{c(4, 4)}.}

\item{max_per_row}{Integer specifying the maximum number of plots per row.
Used to determine the number of columns in multi-panel figures.}
}
\value{
A single character string containing executable Python code.
This code can be passed directly to \code{reticulate::py_run_string()}.
}
\description{
This function generates a block of Python code that, when executed,
produces UMAP density plots for specified metadata variables in an
AnnData object. It is designed for cross-language Seurat → Scanpy
workflows, where an R object is exported to \code{.h5ad} format and
processed in Python using \code{scanpy}.
}
\details{
The generated Python code will:
\enumerate{
\item Import required libraries (\code{scanpy} and \code{matplotlib}).
\item Read the \code{.h5ad} file into an AnnData object.
\item For each metadata variable in \code{density_specs}, compute and plot
UMAP density maps using \code{scanpy.tl.embedding_density()} and
\code{scanpy.pl.embedding_density()}.
\item Save each density plot to a PNG file named
\code{density_<meta_col>.png}.
}
}
\examples{
\dontrun{
density_plot_specs <- list(
  cohort_tp = list(order = c("Chronic Pre-DAA", "Chronic Post-DAA")),
  protein_group = list(order = c("Group1", "Group2"))
)

py_code <- density_plot(
  file_name = "mydata.h5ad",
  density_specs = density_plot_specs
)

}

}
\seealso{
\code{\link[reticulate]{py_run_string}}
}
