% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_vis.R
\name{features_plots}
\alias{features_plots}
\title{Plot Features with Dynamic Layout and Cutoffs}
\usage{
features_plots(
  object,
  features,
  n_col = 4,
  pt.size = 0.8,
  base_width = 1000,
  base_height = 1000,
  percentile = 0.9,
  file_name = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{features}{Character vector of features to plot.}

\item{n_col}{Integer. Number of columns in the plot grid. Default is 4.}

\item{pt.size}{Point size. Default is 0.8.}

\item{base_width}{Width of each plot in pixels. Default is 1000.}

\item{base_height}{Height of each plot in pixels. Default is 1000.}

\item{percentile}{The percentile (between 0 and 1) of feature values to be shown. The higher the percentile, the sharper the feature. Default is 0.9.}

\item{file_name}{Character string for output png name. If NULL, plot will be printed in console but not saved.
Default is NULL.}
}
\value{
NULL. Plots are saved as PNG files in the working directory.
}
\description{
Generates FeaturePlots for selected features from a Seurat object,
adjusting the plot grid dynamically based on the number of features.
Saves both a scaled and an original version of the plot as PNGs.
}
