% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{heatmap_cell_level}
\alias{heatmap_cell_level}
\title{Generate Clustered Heatmap for Seurat Object}
\usage{
heatmap_cell_level(
  object,
  features,
  idents,
  colors,
  split_by_cluster = T,
  cluster_rows = F,
  cluster_columns = F,
  downsample = FALSE,
  file_name = NULL,
  zlim = c(-2, 2),
  width = 6,
  height = NULL,
  fontsize = 3,
  res = 300
)
}
\arguments{
\item{object}{A Seurat object.}

\item{features}{A character vector or a data frame with a column
\code{gene} containing gene names to plot.}

\item{idents}{Character. Column name in \code{meta.data} containing identity.}

\item{colors}{Character vector of colors to use for clusters.}

\item{split_by_cluster}{Logical. If \code{TRUE}, splits the heatmap by cluster (default = TRUE).}

\item{cluster_rows}{Logical. Whether to cluster rows (genes) in the heatmap (default = FALSE).}

\item{cluster_columns}{Logical. Whether to cluster columns (cells) in the heatmap (default = FALSE).}

\item{downsample}{Logical or numeric. If \code{TRUE}, downsamples to the smallest group size.
If numeric, downsamples each group to that number of cells (default = FALSE).}

\item{file_name}{Character. File name for the saved heatmap PNG.
If \code{NULL}, defaults to \code{"<cluster>_heatmap.png"}.}

\item{zlim}{Numeric vector of length 2 specifying limits for Z-score color scale (default = c(-2, 2)).}

\item{width}{Numeric. Width of the output PNG in inches (default = 6).}

\item{height}{Numeric. Height of the output PNG in inches (default = 8).}

\item{fontsize}{fontsize of rownames. Default is 3.}

\item{res}{Numeric. Resolution of the output PNG in DPI (default = 300).}
}
\value{
NULL. Heatmap is saved to file.
}
\description{
Scales data for selected genes, arranges cells by cluster, and produces a heatmap
with custom colors for clusters. Saves the heatmap as a PNG file.
}
