% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{preprocess_obj}
\alias{preprocess_obj}
\title{Preprocess Seurat Object.}
\usage{
preprocess_obj(
  object,
  vf.method = "vst",
  nfeatures = 2000,
  regress.out = NULL,
  model.use = "linear",
  pseudobulk_var = NULL,
  pseudo_cutoff = 30
)
}
\arguments{
\item{object}{A Seurat object.}

\item{vf.method}{Character. Method for variable feature selection. Options: "vst", "dispersion", "mean.var.plot", "pseudobulk", "sct", "anchor".}

\item{nfeatures}{Integer. Number of variable features to select. Default is 2000.}

\item{regress.out}{Character vector. Variables to regress out during scaling. Default is NULL.}

\item{model.use}{Character. Model to use in scaling ("linear" or "poisson"). Default is "linear".}

\item{pseudobulk_var}{Character. Identity variable to use for pseudobulk aggregation. Required if vf.method = "pseudobulk".}

\item{pseudo_cutoff}{Numeric. Minimum cells per identity to include in pseudobulk. Default is 30.}
}
\value{
A Seurat object processed according to the chosen variable feature method.
}
\description{
This function normalizes, selects variable features, and scales a Seurat object
based on the chosen variable feature method and scaling method.
}
